Major refactor, message passing to UDS
rewrote major sections of the codebase to use UDS for message passing.
This allows for the client frontend and the client backend to run
seperately without blocking each other, and make better use of the actix
actor system. Next steps are to figure out the schema for sending data
from the server to the UI, and what data the UI needs.

+++ crates/messages/src/models/header.rs
@@ -131,7 +131,7 @@ fn uint16_to_bytes_big(value: u16) -> [u8; 2] {
     [(value >> 8) as u8, (value & 0xFF) as u8]
 }
 
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub enum PacketFrequency {
     High,
     Medium,
@@ -166,7 +166,7 @@ impl PacketFrequency {
                 // 4 byte ID
                 bytes.push(0xFF);
                 bytes.push(0xFF);
-                let id_bytes = uint16_to_bytes_big(header.id as u16);
+                let id_bytes = uint16_to_bytes_big(header.id);
                 bytes.extend_from_slice(&id_bytes);
             }
             PacketFrequency::Fixed => {
